+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Prefix your message with one of the following:
 - [Security] in case of vulnerabilities.
 -->
 
+## Unreleased
+
+- [Added] Support for `maximum_value`, `minimum_value`, and `cycle` options.
+
 ## v0.2.1 - 2021-12-21
 
 - [Changed] Dump schema names that aren't on the default search path.